+++ packages/gatsby-theme-aoi-blog-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-alpha.1](https://github.com/cieloazul310/gatsby-aoi/compare/v3.0.0-alpha.0...v3.0.0-alpha.1) (2022-11-01)
+
+**Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi-blog-cli
+
 # [3.0.0-alpha.0](https://github.com/cieloazul310/gatsby-aoi/compare/v2.8.0...v3.0.0-alpha.0) (2022-11-01)
 
 **Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi-blog-cli